lazy load for routing, moved model to app folder
