+++ WebStore/src/app/app-routing.module.ts
@@ -1,19 +1,25 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
-import { ProductListComponent } from './product-list/product-list.component';
 import { ProductComponent } from './product/product.component';
-import { ShoppingCartComponent } from './shopping-cart/shopping-cart.component';
 
 
 
 const routes: Routes = [
   { path: '', redirectTo: '/productList', pathMatch: 'full' },
-  { path: 'productList', component: ProductListComponent },
-  { path: 'cart', component: ShoppingCartComponent },
+  {
+    path: 'productList',
+    loadChildren: () => import('./product-list/product-list.module').then(
+      (m) => m.ProductListModule
+    )
+  },
+  {
+    path: 'cart', loadChildren: () => import('./shopping-cart/shopping-cart.module').then(
+      (m) => m.ShoppngCartModule
+    )
+  },
   { path: 'product/:id', component: ProductComponent }
- 
-  
+
+
 ];
 
 @NgModule({
