+++ WebStore/src/app/app.module.ts
@@ -4,22 +4,15 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { ProductComponent } from './product/product.component';
-import { ProductListComponent } from './product-list/product-list.component';
 import { HeaderComponent } from './header/header.component';
-import { ShoppingCartComponent } from './shopping-cart/shopping-cart.component';
 import { HttpClientModule } from '@angular/common/http';
-import { CartTotalComponent } from './cart-total/cart-total.component';
-import { ProductLineComponent } from './product-line/product-line.component';
+
 
 @NgModule({
   declarations: [
     AppComponent,
     ProductComponent,
-    ProductListComponent,
     HeaderComponent,
-    ShoppingCartComponent,
-    CartTotalComponent,
-    ProductLineComponent
   ],
   imports: [
     BrowserModule,
