+++ WebStore/src/app/model/product.ts
