+++ WebStore/src/app/model/rating.ts
