+++ WebStore/src/app/product-list/product-list-routing.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+
+import { ProductListComponent } from './product-list.component';
+
+const routes: Routes = [{ path: '', component: ProductListComponent }];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class ProductListRoutingModule {}
