+++ WebStore/src/app/shopping-cart/shopping-cart-routing.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { ShoppingCartComponent } from './shopping-cart.component';
+
+
+const routes: Routes = [{ path: '', component: ShoppingCartComponent }];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class ShoppingCartRoutingModule {}
